feat: limit operator for WindowedStream
