--- crates/fluxus-api/tests/datastream.rs
@@ -1,20 +1,0 @@
-use fluxus_api::{CollectionSink, CollectionSource, DataStream};
-
-#[test]
-fn test_limit() {
-    tokio_test::block_on(async {
-        let numbers = vec![1, 2, 3, 4, 5];
-        // TransformSourceWithOperator::new();
-        let source = CollectionSource::new(numbers);
-        let sink = CollectionSink::new();
-
-        DataStream::new(source)
-            .limit(2)
-            .sink(sink.clone())
-            .await
-            .unwrap();
-
-        let data = sink.get_data();
-        assert_eq!(data, vec![1, 2]);
-    })
-}
