+++ crates/fluxus-api/tests/integrition_test.rs
@@ -1,1 +1,1 @@
-mod datastream;
+mod datastreams;
